requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
Pillow==10.0.0
pluggy==1.3.0
psycopg2-binary==2.9.7
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
